Rename: 토스트 위치 컴포넌트 폴더로 이동
